refactor (fe): moved add module drawer to its own file
